+++ app/src/main/java/com/pinecreek/game/GameSurface.java
@@ -1,8 +1,10 @@
 package com.pinecreek.game;
 
+import android.annotation.SuppressLint;
 import android.content.Context;
 import android.opengl.GLSurfaceView;
 
+@SuppressLint("ViewConstructor")
 public class GameSurface extends GLSurfaceView {
     public final GameRenderer renderer;
 
@@ -15,3 +17,5 @@ public class GameSurface extends GLSurfaceView {
         setRenderMode(GLSurfaceView.RENDERMODE_CONTINUOUSLY);
     }
 }
+
+[executed on device: festice-virtual-machine (07fc5208-706b-4ca8-850a-ef91db884468)]